+++ build-release/_deps/googlebenchmark-build/CTestTestfile.cmake
@@ -1,7 +1,0 @@
-# CMake generated Testfile for 
-# Source directory: /Users/ericzhou/Desktop/Productivity/Projects/Serious/MatchingEngine/build-release/_deps/googlebenchmark-src
-# Build directory: /Users/ericzhou/Desktop/Productivity/Projects/Serious/MatchingEngine/build-release/_deps/googlebenchmark-build
-# 
-# This file includes the relevant testing commands required for 
-# testing this directory and lists subdirectories to be tested as well.
-subdirs("src")