+++ build-release/_deps/googlebenchmark-build/src/CTestTestfile.cmake
@@ -1,6 +1,0 @@
-# CMake generated Testfile for 
-# Source directory: /Users/ericzhou/Desktop/Productivity/Projects/Serious/MatchingEngine/build-release/_deps/googlebenchmark-src/src
-# Build directory: /Users/ericzhou/Desktop/Productivity/Projects/Serious/MatchingEngine/build-release/_deps/googlebenchmark-build/src
-# 
-# This file includes the relevant testing commands required for 
-# testing this directory and lists subdirectories to be tested as well.